accounts/urls.py
from django.urls import path
from accounts import views
from rest_framework_simplejwt.views import (
    TokenObtainPairView,
    TokenRefreshView,
)

app_name = "accounts"
urlpatterns = [
    path("", views.SignupView.as_view(), name="signup"),
    path("login/", views.LoginView.as_view(), name="login"),
    path("<str:username>/", views.detail),
    path("token/refresh/", TokenRefreshView.as_view(), name="token_refresh"),
]
